revised shock tube case
