+++ tests/2d_examples/test_1d_shock_tube/regression_test_tool/WaveBody_MaximumSpeed_ensemble_averaged_mean_variance.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<meanvariance>
+    <Mean_Element>
+        <Snapshot_0 MaximumSpeed_0="0.94802267920457517" />
+        <Snapshot_100 MaximumSpeed_0="1.5826945203235738" />
+        <Snapshot_200 MaximumSpeed_0="1.6312010060037576" />
+    </Mean_Element>
+    <Variance_Element>
+        <Snapshot_0 MaximumSpeed_0="8.9874700028622079e-05" />
+        <Snapshot_100 MaximumSpeed_0="0.00025049219446622676" />
+        <Snapshot_200 MaximumSpeed_0="0.00026608167219876734" />
+    </Variance_Element>
+</meanvariance>
