+++ tests/2d_examples/test_1d_shock_tube/regression_test_tool/WaveBody_MaximumSpeed_runtimes.dat
@@ -1,0 +1,3 @@
+true
+6
+5
