+++ tests/2d_examples/test_1d_shock_tube/shock_tube.cpp
@@ -4,9 +4,93 @@
  * @details See https://doi.org/10.1016/j.jcp.2010.08.019 for the detailed problem setup.
  * @author 	Zhentong Wang and Xiangyu Hu
  */
-#include "shock_tube.h"
+#include "common_compressible_eulerian_classes.hpp" // eulerian classes for compressible fluid only.
+#include "common_shared_eulerian_classes.h"         // shared eulerian classes for weakly-compressible and compressible fluid.
 #include "sphinxsys.h"
 using namespace SPH;
+//----------------------------------------------------------------------
+//	Basic geometry parameters and numerical setup.
+//----------------------------------------------------------------------
+Real DL = 5.0;                           /**< Tube length. */
+Real particle_spacing_ref = 1.0 / 200.0; /**< Initial reference particle spacing. */
+Real DH = particle_spacing_ref * 4;      /**< Tube height. */
+/** Domain bounds of the system. */
+BoundingBox system_domain_bounds(Vec2d(-2.0 / 5.0 * DL, 0.0), Vec2d(3.0 / 5.0 * DL, DH));
+Real rho0_l = 1.0;              /**< initial density of left state. */
+Real rho0_r = 0.125;            /**< initial density of right state. */
+Vecd velocity_l = Vecd::Zero(); /**< initial velocity of left state. */
+Vecd velocity_r = Vecd::Zero(); /**< initial velocity of right state. */
+Real p_l = 1.0;                 /**< initial pressure of left state. */
+Real p_r = 0.1;                 /**< initial pressure of right state. */
+//----------------------------------------------------------------------
+//	Global parameters on material properties
+//----------------------------------------------------------------------
+Real heat_capacity_ratio = 1.4; /**< heat capacity ratio. */
+//----------------------------------------------------------------------
+//	Cases-dependent geometry
+//----------------------------------------------------------------------
+class WaveBlock : public MultiPolygonShape
+{
+  public:
+    explicit WaveBlock(const std::string &body_name)
+        : MultiPolygonShape(body_name)
+    {
+        std::vector<Vecd> waves_block_shape{
+            Vecd(-2.0 / 5.0 * DL, 0.0), Vecd(-2.0 / 5.0 * DL, DH), Vecd(3.0 / 5.0 * DL, DH),
+            Vecd(3.0 / 5.0 * DL, 0.0), Vecd(-2.0 / 5.0 * DL, 0.0)};
+        multi_polygon_.addAPolygon(waves_block_shape, ShapeBooleanOps::add);
+    }
+};
+//----------------------------------------------------------------------
+//	Case-dependent initial condition.
+//----------------------------------------------------------------------
+class ShockTubeInitialCondition
+    : public fluid_dynamics::FluidInitialCondition
+{
+  public:
+    explicit ShockTubeInitialCondition(SPHBody &sph_body)
+        : FluidInitialCondition(sph_body), pos_(particles_->pos_), vel_(particles_->vel_),
+          rho_(particles_->rho_), p_(*particles_->getVariableByName<Real>("Pressure"))
+    {
+        particles_->registerVariable(mom_, "Momentum");
+        particles_->registerVariable(dmom_dt_, "MomentumChangeRate");
+        particles_->registerVariable(dmom_dt_prior_, "OtherMomentumChangeRate");
+        particles_->registerVariable(E_, "TotalEnergy");
+        particles_->registerVariable(dE_dt_, "TotalEnergyChangeRate");
+        particles_->registerVariable(dE_dt_prior_, "OtherEnergyChangeRate");
+        gamma_ = heat_capacity_ratio;
+    };
+    void update(size_t index_i, Real dt)
+    {
+        if (pos_[index_i][0] < DL / 10.0)
+        {
+            // initial left state pressure,momentum and energy profile
+            rho_[index_i] = rho0_l;
+            p_[index_i] = p_l;
+            Real rho_e = p_[index_i] / (gamma_ - 1.0);
+            vel_[index_i] = velocity_l;
+            mom_[index_i] = rho0_l * velocity_l;
+            E_[index_i] = rho_e + 0.5 * rho_[index_i] * vel_[index_i].squaredNorm();
+        }
+        if (pos_[index_i][0] > DL / 10.0)
+        {
+            // initial right state pressure,momentum and energy profile
+            rho_[index_i] = rho0_r;
+            p_[index_i] = p_r;
+            Real rho_e = p_[index_i] / (gamma_ - 1.0);
+            vel_[index_i] = velocity_r;
+            mom_[index_i] = rho0_r * velocity_r;
+            E_[index_i] = rho_e + 0.5 * rho_[index_i] * vel_[index_i].squaredNorm();
+        }
+    }
+
+  protected:
+    StdLargeVec<Vecd> &pos_, &vel_;
+    StdLargeVec<Real> &rho_, &p_;
+    StdLargeVec<Vecd> mom_, dmom_dt_, dmom_dt_prior_;
+    StdLargeVec<Real> E_, dE_dt_, dE_dt_prior_;
+    Real gamma_;
+};
 //----------------------------------------------------------------------
 //	Main program starts here.
 //----------------------------------------------------------------------
@@ -30,7 +114,6 @@ int main(int ac, char *av[])
     //	The contact relation defines the particle configuration between the bodies.
     //----------------------------------------------------------------------
     InnerRelation wave_body_inner(wave_body);
-    // using correction Matrix to correct kernel gradient.
     InteractionWithUpdate<KernelGradientWithCorrectionInner> kernel_gradient_update(wave_body_inner);
     //----------------------------------------------------------------------
     //	Define the main numerical methods used in the simulation.
@@ -39,13 +122,9 @@ int main(int ac, char *av[])
     SimpleDynamics<ShockTubeInitialCondition> waves_initial_condition(wave_body);
     wave_body.addBodyStateForRecording<Real>("TotalEnergy");
     wave_body.addBodyStateForRecording<Real>("Density");
-    // Initialize particle acceleration.
     SimpleDynamics<EulerianCompressibleTimeStepInitialization> initialize_wave_step(wave_body);
-    // Periodic BCs in y direction.
     PeriodicConditionUsingCellLinkedList periodic_condition_y(wave_body, wave_body.getBodyShapeBounds(), yAxis);
-    // Time step size with considering sound wave speed.
     ReduceDynamics<EulerianCompressibleAcousticTimeStepSize> get_wave_time_step_size(wave_body);
-    // Pressure, density and energy relaxation algorithm by use HLLC Riemann solver.
     InteractionWithUpdate<Integration1stHalfHLLCRiemann> pressure_relaxation(wave_body_inner);
     InteractionWithUpdate<Integration2ndHalfHLLCRiemann> density_and_energy_relaxation(wave_body_inner);
     //----------------------------------------------------------------------
@@ -53,6 +132,8 @@ int main(int ac, char *av[])
     //	Regression tests are also defined here.
     //----------------------------------------------------------------------
     BodyStatesRecordingToPlt body_states_recording(io_environment, sph_system.real_bodies_);
+    RegressionTestEnsembleAverage<ReducedQuantityRecording<ReduceDynamics<MaximumSpeed>>>
+        write_maximum_speed(io_environment, wave_body);
     //----------------------------------------------------------------------
     //	Prepare the simulation with cell linked list, configuration
     //	and case specified initial condition if necessary.
@@ -97,6 +178,7 @@ int main(int ac, char *av[])
 
             if (number_of_iterations % screen_output_interval == 0)
             {
+                write_maximum_speed.writeToFile(number_of_iterations);
                 std::cout << std::fixed << std::setprecision(9) << "N=" << number_of_iterations << "	Time = "
                           << GlobalStaticVariables::physical_time_
                           << "	dt = " << dt << "\n";
@@ -114,6 +196,15 @@ int main(int ac, char *av[])
     TimeInterval tt;
     tt = t4 - t1 - interval;
     std::cout << "Total wall time for computation: " << tt.seconds() << " seconds." << std::endl;
+
+    if (sph_system.GenerateRegressionData())
+    {
+        write_maximum_speed.generateDataBase(1.0e-3, 1.0e-3);
+    }
+    else
+    {
+        write_maximum_speed.testResult();
+    }
 
     return 0;
 }
