--- tests/2d_examples/test_1d_shock_tube/src/CMakeLists.txt
@@ -1,13 +1,0 @@
-set(CMAKE_MODULE_PATH ${CMAKE_MODULE_PATH} ${SPHINXSYS_PROJECT_DIR}/cmake) # main (top) cmake dir
-
-set(CMAKE_VERBOSE_MAKEFILE on)
-
-set(EXECUTABLE_OUTPUT_PATH "${PROJECT_BINARY_DIR}/bin/")
-set(LIBRARY_OUTPUT_PATH ${PROJECT_BINARY_DIR}/lib)
-
-aux_source_directory(. DIR_SRCS)
-add_executable(${PROJECT_NAME} ${DIR_SRCS})
-target_link_libraries(${PROJECT_NAME} extra_sources_2d)
-
-add_test(NAME ${PROJECT_NAME} COMMAND ${PROJECT_NAME} --state_recording=${TEST_STATE_RECORDING})
-set_tests_properties(${PROJECT_NAME} PROPERTIES LABELS "periodic boundary, Eulerian")
